app/requirements.txt
fastapi 
pymongo
uvicorn
passlib[bcrypt]
pydantic[email]
python-multipart
python-jose[cryptography]
